// src/api/dtos/characters-dto.ts
/** Characters dto interface */
export interface CharactersDto {
  /** Name */
  name: string;
  /** Birth year */
  birth_year: string;
  /** Eyes color */
  eye_color: string;
  /** Gender */
  gender: string;
  /** Hair color */
  hair_color: string;
  /** Height */
  height: number;
  /** Home */
  homeworld: string;
  /** Mass */
  mass: string;
  /** Skin color */
  skin_color: string;
}